assignment-1/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
protobuf==3.13.0
PyJWT==1.7.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3
